Simplify TextField to fully controlled also with validation
This is kind fo annoying, because there is no way to
know if the input is valid, except for the red ring
around the TextField the user sees.

A more sophisticated API would be a return value which
holds both the raw string the user has entered as well
as the validated value in whatever type the validation
function returns or an error message.